change app theme to Material
