+++ app/src/main/java/com/aghourservices/ui/fragment/BaseFragment.kt
@@ -9,16 +9,20 @@ import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import com.aghourservices.R
 import com.aghourservices.utils.helper.Event
-import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.google.android.material.bottomappbar.BottomAppBar
+import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.snackbar.Snackbar
 
 open class BaseFragment : Fragment() {
-    private lateinit var bottomNavigationView: BottomNavigationView
+    private lateinit var bottomNavigationView: BottomAppBar
+    private lateinit var fabButton: FloatingActionButton
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         Event.sendScreenName(this::class.simpleName.toString())
-        bottomNavigationView = activity?.findViewById(R.id.bottomView) as BottomNavigationView
+        bottomNavigationView = activity?.findViewById(R.id.bottomAppBar) as BottomAppBar
+        fabButton = activity?.findViewById(R.id.floatingActionButton) as FloatingActionButton
+
     }
 
     fun showToolbar() {
@@ -33,10 +37,12 @@ open class BaseFragment : Fragment() {
 
     fun showBottomNavigation() {
         bottomNavigationView.isVisible = true
+        fabButton.show()
     }
 
     fun hideBottomNavigation() {
         bottomNavigationView.isVisible = false
+        fabButton.hide()
     }
 
     fun onSNACK(view: View, message: String) {
