+++ app/src/main/res/layout/firms_shimmer.xml
@@ -8,7 +8,7 @@
     android:orientation="vertical"
     android:padding="18dp">
 
-    <TextView
+    <androidx.appcompat.widget.AppCompatTextView
         android:layout_width="match_parent"
         android:layout_height="10dp"
         android:layout_gravity="end"
@@ -26,7 +26,7 @@
         android:orientation="horizontal">
 
 
-        <TextView
+        <androidx.appcompat.widget.AppCompatTextView
             android:layout_width="0dp"
             android:layout_height="10dp"
             android:layout_gravity="end"
@@ -36,7 +36,7 @@
 
     </LinearLayout>
 
-    <Button
+    <androidx.appcompat.widget.AppCompatButton
         android:id="@+id/btnCall"
         android:layout_width="match_parent"
         android:layout_height="40dp"
