+++ app/src/main/res/layout/news_shimmer.xml
@@ -14,7 +14,7 @@
         android:orientation="vertical"
         android:padding="18dp">
 
-        <TextView
+        <androidx.appcompat.widget.AppCompatTextView
             android:layout_width="match_parent"
             android:layout_height="10dp"
             android:layout_gravity="end"
@@ -31,7 +31,7 @@
             android:gravity="end"
             android:orientation="horizontal">
 
-            <TextView
+            <androidx.appcompat.widget.AppCompatTextView
                 android:layout_width="0dp"
                 android:layout_height="10dp"
                 android:layout_gravity="end"
