+++ app/src/main/res/values-night/colors.xml
@@ -111,5 +111,6 @@
     <color name="tagsColor">#FFFFFF</color>
     <color name="edtAddArticle">#242D33</color>
     <color name="strokeAddArticle">#2B3740</color>
+    <color name="settingIconColor">#FFFFFF</color>
 
 </resources>
