+++ app/src/main/res/values-night/themes.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources xmlns:tools="http://schemas.android.com/tools">
 
-    <style name="Theme.DarkApp" parent="Theme.AppCompat.DayNight.NoActionBar">
+    <style name="Theme.DarkApp" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Status bar color. -->
         <item name="android:statusBarColor">@color/toolbarBg</item>
         <item name="colorPrimary">@color/itemNavColor</item>
@@ -27,12 +27,12 @@
 
     <style name="CategoriesTheme" parent="Theme.DarkApp.NoActionBar" />
 
-    <style name="AddDataTheme" parent="Theme.AppCompat.DayNight.NoActionBar">
+    <style name="AddDataTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <item name="android:statusBarColor">@color/toolbarBg</item>
         <item name="cardBackgroundColor">@color/createOrSignInAccountBg</item>
     </style>
 
-    <style name="SplashScreenTheme" parent="Theme.AppCompat.DayNight.NoActionBar">
+    <style name="SplashScreenTheme" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <item name="android:statusBarColor">@color/splashScreenBg</item>
         <item name="windowActionBar">true</item>
         <item name="android:windowBackground">@color/splashScreenBg</item>
@@ -40,12 +40,12 @@
         <item name="android:windowLightNavigationBar" tools:targetApi="o_mr1">true</item>
     </style>
 
-    <style name="AlertDialogTheme" parent="Theme.AppCompat.DayNight.Dialog.Alert">
+    <style name="AlertDialogTheme" parent="Theme.MaterialComponents.DayNight.Dialog.Alert">
         <item name="fontFamily">@font/segoe_ui_bold</item>
         <item name="android:textSize">14sp</item>
     </style>
 
-    <style name="my_dialog" parent="Theme.AppCompat.Dialog.Alert">
+    <style name="my_dialog" parent="Theme.MaterialComponents.Dialog.Alert">
         <item name="android:windowMinWidthMajor">0dp</item>
         <item name="android:windowMinWidthMinor">0dp</item>
         <item name="android:background">@drawable/news_card_ripple</item>
