+++ app/src/main/res/values/themes.xml
@@ -1,11 +1,14 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application light theme. -->
-    <style name="Theme.LightApp" parent="Theme.AppCompat.Light.NoActionBar">
+    <style name="Theme.LightApp" parent="Theme.MaterialComponents.Light.NoActionBar">
         <!-- Status bar color. -->
         <item name="android:statusBarColor">@color/toolbarBg</item>
+
         <!-- Customize your theme here. -->
         <item name="android:navigationBarColor">@color/navDrawerBg</item>
         <item name="android:windowLightNavigationBar" tools:targetApi="o_mr1">true</item>
+        <item name="android:windowLightStatusBar" tools:targetApi="m">true</item>
+
         <item name="drawerArrowStyle">@style/drawerToggle</item>
     </style>
 
@@ -14,21 +17,17 @@
         <item name="windowNoTitle">true</item>
     </style>
 
-    <!--Custom the text on Bottom Navigation-->
-    <style name="BottomNavigationTextStyle">
-        <item name="android:textSize">14sp</item>
-        <item name="android:textStyle">bold</item>
-        <item name="fontFamily">@font/segoe_ui_bold</item>
+    <style name="CategoriesTheme" parent="Theme.LightApp.NoActionBar" >
+        <item name="android:windowLightStatusBar" tools:targetApi="m">true</item>
     </style>
 
-    <style name="CategoriesTheme" parent="Theme.LightApp.NoActionBar" />
-
-    <style name="AddDataTheme" parent="Theme.AppCompat.Light.NoActionBar">
+    <style name="AddDataTheme" parent="Theme.MaterialComponents.Light.NoActionBar">
         <item name="android:statusBarColor">@color/toolbarBg</item>
         <item name="cardBackgroundColor">@color/createOrSignInAccountBg</item>
+        <item name="android:windowLightStatusBar" tools:targetApi="m">true</item>
     </style>
 
-    <style name="SplashScreenTheme" parent="Theme.AppCompat.Light.NoActionBar">
+    <style name="SplashScreenTheme" parent="Theme.MaterialComponents.Light.NoActionBar">
         <item name="android:statusBarColor">@color/splashScreenBg</item>
         <item name="windowActionBar">true</item>
         <item name="android:windowBackground">@color/splashScreenBg</item>
@@ -36,40 +35,13 @@
         <item name="android:windowLightNavigationBar" tools:targetApi="o_mr1">true</item>
     </style>
 
-    <style name="AlertDialogTheme" parent="Theme.AppCompat.Light.Dialog.Alert">
+    <style name="AlertDialogTheme" parent="Theme.MaterialComponents.Light.Dialog.Alert">
         <item name="fontFamily">@font/segoe_ui_bold</item>
         <item name="android:textSize">14sp</item>
     </style>
 
-    <style name="my_dialog" parent="Theme.AppCompat.Dialog.Alert">
-        <item name="android:windowMinWidthMajor">0dp</item>
-        <item name="android:windowMinWidthMinor">0dp</item>
-        <item name="android:background">@drawable/news_card_ripple</item>
-        <item name="android:textColor">@color/dialogText</item>
-    </style>
-
     <style name="drawerToggle" parent="Widget.AppCompat.DrawerArrowToggle">
         <item name="color">@color/drawerToggle</item>
-    </style>
-    <!--All Text Font Style-->
-    <style name="RobotoTextViewStyle" parent="android:Widget.TextView">
-        <item name="android:fontFamily">@font/cairo_black</item>
-    </style>
-
-    <style name="SegoeTextRegular" parent="android:Widget.TextView">
-        <item name="fontFamily">@font/segoe_ui_regular</item>
-        <item name="android:textStyle">bold</item>
-        <item name="android:autoSizeMaxTextSize" tools:targetApi="o">25sp</item>
-    </style>
-
-    <style name="SegoeTextBold" parent="android:Widget.TextView">
-        <item name="fontFamily">@font/segoe_ui_bold</item>
-        <item name="android:textSize">11sp</item>
-        <item name="android:autoSizeMaxTextSize" tools:targetApi="o">25sp</item>
-    </style>
-    <!--All Text Buttons Font Style-->
-    <style name="RobotoButtonStyle" parent="android:Widget.Holo.Button">
-        <item name="android:fontFamily">@font/cairo_black</item>
     </style>
 
 </resources>
